refactor: move codegen functions to core
