+++ core/src/main/java/io/sundr/functions/StringReplace.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package io.sundr.codegen.functions;
+package io.sundr.functions;
 
 import java.util.function.Function;
 import java.util.regex.Matcher;
