+++ core/src/test/java/io/sundr/functions/SingularizeTest.java
@@ -14,22 +14,22 @@
  *    limitations under the License.
  */
 
-package io.sundr.codegen.functions;
+package io.sundr.functions;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
 
-public class PluralizeTest {
+public class SingularizeTest {
 
   @Test
-  public void testPluralize() {
+  public void testSingularize() {
 
-    assertEquals("addresses", Pluralize.FUNCTION.apply("address"));
-    assertEquals("numbers", Pluralize.FUNCTION.apply("number"));
-    assertEquals("vertices", Pluralize.FUNCTION.apply("vertex"));
-    assertEquals("endpoints", Pluralize.FUNCTION.apply("endpoints"));
-    assertEquals("ingresses", Pluralize.FUNCTION.apply("ingress"));
+    assertEquals("address", Singularize.FUNCTION.apply("addresses"));
+    assertEquals("ADDRESS", Singularize.FUNCTION.apply("ADDRESSES"));
+
+    assertEquals("number", Singularize.FUNCTION.apply("numbers"));
+    assertEquals("NUMBER", Singularize.FUNCTION.apply("NUMBERS"));
   }
 
 }
